refactor network default implementation
