+++ CommonWallet/Core/Classes/Common/Fee/FeeCalculationResult.swift
@@ -1,3 +1,8 @@
+/**
+* Copyright Soramitsu Co., Ltd. All Rights Reserved.
+* SPDX-License-Identifier: GPL-3.0
+*/
+
 import Foundation
 
 public struct FeeCalculationResult {
